Organize everything under lib
