+++ config.ru
@@ -1,2 +1,5 @@
-require './app'
-run Sinatra::Application
+$LOAD_PATH.unshift(File.expand_path("../lib", __FILE__))
+
+require 'board_game_geek'
+
+run BoardGameGeek::App
