+++ lib/board_game_geek.rb
@@ -1,0 +1,8 @@
+require 'super_struct'
+require 'open-uri'
+require 'nokogiri'
+
+module BoardGameGeek
+  autoload :App, 'board_game_geek/app'
+  autoload :Game, 'board_game_geek/game'
+end
