+++ lib/board_game_geek/app/views/about.haml
