+++ lib/board_game_geek/app/views/game.haml
