+++ lib/board_game_geek/app/views/gamesearch.haml
