+++ lib/board_game_geek/app/views/hot.haml
