+++ lib/board_game_geek/app/views/index.haml
