+++ lib/board_game_geek/app/views/searchresults.haml
