+++ lib/board_game_geek/app/views/user.haml
