+++ lib/board_game_geek/app/views/usersearch.haml
